requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
DateTime==5.5
Django==5.1.4
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pathlib==1.0.1
pillow==11.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
typing_extensions==4.12.2
whitenoise==6.8.2
zope.interface==7.2
